+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `__repr__` overloaded for `enums.Enum` which just returns `enums.Enum.__str__`.
 - `Profile.collect()` method which fetch and collect all characters at once.
-- Implemented `aiobungie.crate.fireteams` objects and `Client.fetch_fireteam()` method.
+- Implemented `aiobungie.crate.fireteams` objects and its methods.
 
 ### Changed
 - `RESTClient._request` now takes a string or `rest.RequestMethod` enum for the method.